Fix logs not showing up on all servers (#472)
For some mysterious reason, a traversed object using .map is not formatted by Winston, resulting in the logs being swallowed. For further context, we moved to .map because it's supposed to clone the node before changing it to fix a mutation issue on the API responses (see #469).

This commit goes back to .forEach but ensures nodes that are classes extending from Error are cloned. It also cleans up the noise in the test file by reducing the input samples to a bare minimum.